Added login endpoint
